chore: cleanup code
