+++ src/components/Feed/ActivityComment/ActivityComment.jsx
@@ -2,34 +2,12 @@ import React, { useState } from 'react'
 import * as Styled from './ActivityComment.styled'
 import ActivityHeader from '../ActivityHeader/ActivityHeader'
 import ReactMarkdown from 'react-markdown'
-import ActivityReference from '../ActivityReference/ActivityReference'
 import CommentWrapper from './CommentWrapper'
 import remarkGfm from 'remark-gfm'
-import ActivityCheckbox from '../ActivityCheckbox/ActivityCheckbox'
 import { classNames } from 'primereact/utils'
 import { useSelector } from 'react-redux'
 import CommentInput from '/src/components/CommentInput/CommentInput'
-
-const allowedRefTypes = [
-  'user',
-  'task',
-  'folder',
-  'version',
-  'representation',
-  'workfile',
-  'product',
-]
-const sanitizeURL = (url = '') => {
-  // ensure that the url is valid https url
-  // or a valid {type}:{id} reference
-  if (url.startsWith('https://')) return { url, type: 'url' }
-  if (url.includes(':')) {
-    const sections = url.split(':')
-    const [type, id] = sections
-    if (allowedRefTypes.includes(type) && id && sections.length === 2) return { type, id }
-  }
-  return {}
-}
+import { aTag, inputTag } from './activityMarkdownComponents'
 
 const ActivityComment = ({
   activity = {},
@@ -108,45 +86,10 @@ const ActivityComment = ({
               remarkPlugins={[remarkGfm]}
               urlTransform={(url) => url}
               components={{
-                a: ({ children, href }) => {
-                  const { url, type, id } = sanitizeURL(href)
-
-                  // return regular url
-                  if (url) <a href={href}>{children}</a>
-                  // if no reference type, return regular link with no href
-                  if (!type || !id) return <a>{children}</a>
-
-                  const label = children && children.replace('@', '')
-                  // is this ref the same as the current task id
-                  const isEntity = id === entityId
-
-                  return (
-                    <ActivityReference
-                      name={id}
-                      {...{ type, id, label, projectName, projectInfo }}
-                      variant={isEntity ? 'filled' : 'primary'}
-                      onClick={() =>
-                        type !== 'user' &&
-                        onReferenceClick({ entityId: id, entityType: type, projectName })
-                      }
-                    >
-                      {label}
-                    </ActivityReference>
-                  )
-                },
+                // a links
+                a: (props) => aTag(props, { entityId, projectName, projectInfo, onReferenceClick }),
                 // checkbox inputs
-                input: ({ type, checked, ...props }) => {
-                  if (type === 'checkbox') {
-                    return (
-                      <ActivityCheckbox
-                        checked={checked}
-                        onChange={(e) => onCheckChange && onCheckChange(e, activity)}
-                      />
-                    )
-                  } else {
-                    return <input type={type} disabled {...props} />
-                  }
-                },
+                input: (props) => inputTag(props, { activity, onCheckChange }),
               }}
             >
               {body}
